+++ rust/routee-compass-powertrain/src/routee/mod.rs
@@ -2,4 +2,4 @@ pub mod energy_model_ops;
 pub mod energy_model_service;
 pub mod energy_traversal_model;
 pub mod prediction;
-pub mod vehicles;
+pub mod vehicle;
