+++ rust/routee-compass-powertrain/src/routee/vehicle/default/mod.rs
@@ -1,0 +1,2 @@
+pub mod single_fuel_vehicle;
+pub mod dual_fuel_vehicle;
