+++ rust/routee-compass-powertrain/src/routee/vehicle/mod.rs
@@ -1,0 +1,6 @@
+pub mod default;
+pub mod vehicle_type;
+pub mod vehicle_energy_result;
+
+pub use vehicle_type::{VehicleType, VehicleState};
+pub use vehicle_energy_result::VehicleEnergyResult;
