--- rust/routee-compass-powertrain/src/routee/vehicles/default/mod.rs
@@ -1,2 +1,0 @@
-pub mod conventional;
-pub mod plug_in_hybrid;
