--- rust/routee-compass-powertrain/src/routee/vehicles/mod.rs
@@ -1,4 +1,0 @@
-pub mod default;
-pub mod vehicle_trait;
-
-pub use vehicle_trait::{Vehicle, VehicleEnergyResult, VehicleState};
